requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-auth==1.12.0
hubspot==0.1.14.dev0
hubspot3==3.2.48
idna==2.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.12.0
requests==2.23.0
rsa==4.0
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
